+++ docs/user/changes.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 - feat: support for pChargeInfoSupportPolicy on :class:`trunk details <wxc_sdk.telephony.prem_pstn.trunk.TrunkDetail>` and in :meth:`api.telephony.prem_pstn.trunk.create <wxc_sdk.telephony.prem_pstn.trunk.TrunkApi.create>` and :meth:`api.telephony.prem_pstn.trunk.update <wxc_sdk.telephony.prem_pstn.trunk.TrunkApi.update>`
+- feat: new API :attr:`api.me.go_override <wxc_sdk.me.MeSettingsApi.go_override>`
 
 1.26.0
 ------